VOC数据集裁剪图片做分类来用

+++ jade/ReadVocData.py
@@ -70,7 +70,7 @@ def ProcessXml(xml_path):
             truncated.append('0')
 
         bbox = obj.find('bndbox')
-        bboxes.append((int(bbox.find('xmin').text) / int(shape[1]) ,int(bbox.find('ymin').text) / int(shape[0]),int(bbox.find('xmax').text)/int(shape[1]),int(bbox.find('ymax').text)/int(shape[0])))
+        bboxes.append((float(bbox.find('xmin').text) / float(shape[1]) ,float(bbox.find('ymin').text) / float(shape[0]),float(bbox.find('xmax').text)/float(shape[1]),float(bbox.find('ymax').text)/float(shape[0])))
     imagename = GetLastDir(xml_path)[:-4]+'.jpg'
     return imagename,shape, bboxes, labels_text,labels, difficult, truncated
